+++ docs/CHANGELOG.md
@@ -13,10 +13,15 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
-## TBD
+## [0.99.56] - 24-12-01
 ### Fixed
 - Views used for edges that implements other views now correctly include all properties in the generated SDK.
 - The `.aggregate("count")` method now correctly returns a single value instead of a list of values.
+- If a view as a property `classes` the generated SDK no longer raises a `SyntaxError`.
+- Using `.earliest()/.latest()` no longer returns multiple items when used as part nested query.
+
+### Improved
+- Better error message for typos in the `.select()` method.
 
 ## [0.99.55] - 24-11-29